pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["data", "mesh", "agents"]
package-dir = {"" = "src"}

[project]
name = "ai-agents"
version = "0.1.0"
description = "A project with two AI agents communicating together"
authors = [
    { name="Your Name", email="your.email@example.com" }
]
dependencies = [
    "flask",
    "openai",
    "pytest",
    "python-dotenv",
    "streamlit"
]
